Final commit probably
I switched from XMonad to i3 as a temporary step while I eventually
switch to Sway, which has the same configuration as i3.

I might revisit this in the future if I'm greatly unsatisfied by i3, but
i3 natively supports transient states, which would take a decent amount
of time to implement here so in the interest of time, I'm going to leave
this here.